Version 1.1 - Working but broken
This is a working version of the bot. It will run - the commands need regression testing, I tried some of them and they aren't working - whereas some of the other ones are. I will regression test, then fix each command to get every command working - then redeploy the bot in a Beta version.